@@ -1,5 +1,31 @@
-module github.com/photostorm/gatt
+module github.com/xaionaro-go/gatt
 
-go 1.13
+go 1.24.1
 
-require golang.org/x/sys v0.31.0
+require (
+	github.com/facebookincubator/go-belt v0.0.0-20250308011339-62fb7027b11f
+	github.com/xaionaro-go/ctxflow v0.0.0-20250530224412-dee18b630ad7
+	golang.org/x/sys v0.33.0
+)
+
+require (
+	github.com/DataDog/gostackparse v0.7.0 // indirect
+	github.com/go-ng/slices v0.0.0-20230703171042-6195d35636a2 // indirect
+	github.com/go-ng/sort v0.0.0-20220617173827-2cc7cd04f7c7 // indirect
+	github.com/go-ng/xatomic v0.0.0-20230519181013-85c0ec87e55f // indirect
+	github.com/go-ng/xsort v0.0.0-20220617174223-1d146907bccc // indirect
+	github.com/google/uuid v1.6.0 // indirect
+	github.com/huandu/go-tls v0.0.0-20200109070953-6f75fb441850 // indirect
+	github.com/klauspost/cpuid/v2 v2.2.7 // indirect
+	github.com/phuslu/goid v1.0.1 // indirect
+	github.com/sirupsen/logrus v1.9.3 // indirect
+	github.com/xaionaro-go/gorex v0.0.0-20241010205749-bcd59d639c4d // indirect
+	github.com/xaionaro-go/logrustash v0.0.0-20240804141650-d48034780a5f // indirect
+	github.com/xaionaro-go/object v0.0.0-20241026212449-753ce10ec94c // indirect
+	github.com/xaionaro-go/observability v0.0.0-20250525153415-e6c2d935ab34 // indirect
+	github.com/xaionaro-go/spinlock v0.0.0-20200518175509-30e6d1ce68a1 // indirect
+	github.com/xaionaro-go/unsafetools v0.0.0-20241024014258-a46e1ce3763e // indirect
+	github.com/xaionaro-go/xsync v0.0.0-20250511184922-deec5fb01a0f // indirect
+	golang.org/x/exp v0.0.0-20240823005443-9b4947da3948 // indirect
+	lukechampine.com/blake3 v1.3.0 // indirect
+)
