@@ -1,17 +1,20 @@
-// +build
+//go:build ignore
+// +build ignore
 
 package main
 
 import (
 	"bytes"
+	"context"
 	"flag"
 	"fmt"
-	"log"
+
 	"time"
 
-	"github.com/photostorm/gatt"
-	"github.com/photostorm/gatt/examples/service"
-	"github.com/photostorm/gatt/linux/cmd"
+	"github.com/facebookincubator/go-belt/tool/logger"
+	"github.com/xaionaro-go/gatt"
+	"github.com/xaionaro-go/gatt/examples/service"
+	"github.com/xaionaro-go/gatt/linux/cmd"
 )
 
 // server_lnx implements a GATT server.
@@ -27,29 +30,31 @@ var (
 	chk   = flag.Bool("chk", true, "Check device LE support")
 )
 
-// cmdReadBDAddr implements cmd.CmdParam for demostrating LnxSendHCIRawCommand()
+// cmdReadBDAddr implements cmd.CmdParam for demonstrating LnxSendHCIRawCommand()
 type cmdReadBDAddr struct{}
 
 func (c cmdReadBDAddr) Marshal(b []byte) {}
 func (c cmdReadBDAddr) Opcode() int      { return 0x1009 }
 func (c cmdReadBDAddr) Len() int         { return 0 }
 
-// Get bdaddr with LnxSendHCIRawCommand() for demo purpose
-func bdaddr(d gatt.Device) {
-	rsp := bytes.NewBuffer(nil)
-	if err := d.Option(gatt.LnxSendHCIRawCommand(&cmdReadBDAddr{}, rsp)); err != nil {
+// Get bdAddr with LnxSendHCIRawCommand() for demo purpose
+func bdAddr(ctx context.Context, d gatt.Device) {
+	resp := bytes.NewBuffer(nil)
+	if err := d.Option(gatt.LnxSendHCIRawCommand(ctx, &cmdReadBDAddr{}, resp)); err != nil {
 		fmt.Printf("Failed to send HCI raw command, err: %s", err)
 	}
-	b := rsp.Bytes()
+	b := resp.Bytes()
 	if b[0] != 0 {
 		fmt.Printf("Failed to get bdaddr with HCI Raw command, status: %d", b[0])
 	}
-	log.Printf("BD Addr: %02X:%02X:%02X:%02X:%02X:%02X", b[6], b[5], b[4], b[3], b[2], b[1])
+	logger.Debugf(ctx, "BD Addr: %02X:%02X:%02X:%02X:%02X:%02X", b[6], b[5], b[4], b[3], b[2], b[1])
 }
 
 func main() {
+	ctx := context.Background()
 	flag.Parse()
 	d, err := gatt.NewDevice(
+		ctx,
 		gatt.LnxMaxConnections(*mc),
 		gatt.LnxDeviceID(*dev, *chk),
 		gatt.LnxSetAdvertisingParameters(&cmd.LESetAdvertisingParameters{
@@ -60,40 +65,44 @@ func main() {
 	)
 
 	if err != nil {
-		log.Printf("Failed to open device, err: %s", err)
-		return
+		logger.Fatalf(ctx, "failed to open device, err: %s", err)
 	}
 
 	// Register optional handlers.
 	d.Handle(
-		gatt.CentralConnected(func(c gatt.Central) { log.Println("Connect: ", c.ID()) }),
-		gatt.CentralDisconnected(func(c gatt.Central) { log.Println("Disconnect: ", c.ID()) }),
+		ctx,
+		gatt.CentralConnected(func(ctx context.Context, c gatt.Central) {
+			fmt.Println("Connect: ", c.ID())
+		}),
+		gatt.CentralDisconnected(func(ctx context.Context, c gatt.Central) {
+			fmt.Println("Disconnect: ", c.ID())
+		}),
 	)
 
 	// A mandatory handler for monitoring device state.
-	onStateChanged := func(d gatt.Device, s gatt.State) {
+	onStateChanged := func(ctx context.Context, d gatt.Device, s gatt.State) {
 		fmt.Printf("State: %s\n", s)
 		switch s {
 		case gatt.StatePoweredOn:
 			// Get bdaddr with LnxSendHCIRawCommand()
-			bdaddr(d)
+			bdAddr(ctx, d)
 
 			// Setup GAP and GATT services.
-			d.AddService(service.NewGapService(*name))
-			d.AddService(service.NewGattService())
+			d.AddService(ctx, service.NewGapService(*name))
+			d.AddService(ctx, service.NewGattService())
 
 			// Add a simple counter service.
 			s1 := service.NewCountService()
-			d.AddService(s1)
+			d.AddService(ctx, s1)
 
 			// Add a simple counter service.
 			s2 := service.NewBatteryService()
-			d.AddService(s2)
+			d.AddService(ctx, s2)
 			uuids := []gatt.UUID{s1.UUID(), s2.UUID()}
 
 			// If id is zero, advertise name and services statically.
 			if *id == time.Duration(0) {
-				d.AdvertiseNameAndServices(*name, uuids)
+				d.AdvertiseNameAndServices(ctx, *name, uuids)
 				break
 			}
 
@@ -101,11 +110,11 @@ func main() {
 			go func() {
 				for {
 					// Advertise as a RedBear Labs iBeacon.
-					d.AdvertiseIBeacon(gatt.MustParseUUID("5AFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"), 1, 2, -59)
+					d.AdvertiseIBeacon(ctx, gatt.MustParseUUID("5AFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF"), 1, 2, -59)
 					time.Sleep(*id)
 
 					// Advertise name and services.
-					d.AdvertiseNameAndServices(*name, uuids)
+					d.AdvertiseNameAndServices(ctx, *name, uuids)
 					time.Sleep(*ii)
 				}
 			}()
@@ -114,6 +123,6 @@ func main() {
 		}
 	}
 
-	d.Init(onStateChanged)
-	select {}
+	d.Start(ctx, onStateChanged)
+	<-ctx.Done()
 }
