@@ -1,16 +1,18 @@
 package linux
 
 import (
+	"context"
 	"fmt"
 	"io"
-	"log"
 	"sync"
 
 	"golang.org/x/sys/unix"
 
-	"github.com/photostorm/gatt/linux/cmd"
-	"github.com/photostorm/gatt/linux/evt"
-	"github.com/photostorm/gatt/linux/util"
+	"github.com/facebookincubator/go-belt/tool/logger"
+	"github.com/xaionaro-go/gatt/linux/cmd"
+	"github.com/xaionaro-go/gatt/linux/consts"
+	"github.com/xaionaro-go/gatt/linux/event"
+	"github.com/xaionaro-go/gatt/linux/util"
 )
 
 type HCI struct {
@@ -18,12 +20,12 @@ type HCI struct {
 	AcceptSlaveHandler   func(pd *PlatData)
 	AdvertisementHandler func(pd *PlatData)
 
-	d io.ReadWriteCloser
-	c *cmd.Cmd
-	e *evt.Evt
+	device io.ReadWriteCloser
+	cmd    *cmd.Cmd
+	event  *event.Event
 
-	plist   map[bdaddr]*PlatData
-	plistmu *sync.Mutex
+	plist      map[bdaddr]*PlatData
+	plistMutex *sync.Mutex
 
 	bufCnt  chan struct{}
 	bufSize int
@@ -31,12 +33,12 @@ type HCI struct {
 	pool     *util.BytePool
 	loopDone chan bool
 
-	maxConn int
-	connsmu *sync.Mutex
-	conns   map[uint16]*conn
+	maxConn    int
+	connsMutex *sync.Mutex
+	conns      map[uint16]*conn
 
-	adv   bool
-	advmu *sync.Mutex
+	adv      bool
+	advMutex *sync.Mutex
 }
 
 type bdaddr [6]byte
@@ -52,21 +54,21 @@ type PlatData struct {
 	Conn io.ReadWriteCloser
 }
 
-func NewHCI(devID int, chk bool, maxConn int) (*HCI, error) {
-	d, err := newDevice(devID, chk)
+func NewHCI(ctx context.Context, devID int, chk bool, maxConn int) (*HCI, error) {
+	device, err := newDevice(ctx, devID, chk)
 	if err != nil {
 		return nil, err
 	}
-	c := cmd.NewCmd(d)
-	e := evt.NewEvt()
+	cmd := cmd.NewCmd(ctx, device)
+	ev := event.NewEvent()
 
 	h := &HCI{
-		d: d,
-		c: c,
-		e: e,
+		device: device,
+		cmd:    cmd,
+		event:  ev,
 
-		plist:   make(map[bdaddr]*PlatData),
-		plistmu: &sync.Mutex{},
+		plist:      make(map[bdaddr]*PlatData),
+		plistMutex: &sync.Mutex{},
 
 		bufCnt:  make(chan struct{}, 15-1),
 		bufSize: 27,
@@ -74,89 +76,89 @@ func NewHCI(devID int, chk bool, maxConn int) (*HCI, error) {
 		pool:     util.NewBytePool(4096, 16),
 		loopDone: make(chan bool),
 
-		maxConn: maxConn,
-		connsmu: &sync.Mutex{},
-		conns:   map[uint16]*conn{},
+		maxConn:    maxConn,
+		connsMutex: &sync.Mutex{},
+		conns:      map[uint16]*conn{},
 
-		advmu: &sync.Mutex{},
+		advMutex: &sync.Mutex{},
 	}
 
-	e.HandleEvent(evt.LEMeta, evt.HandlerFunc(h.handleLEMeta))
-	e.HandleEvent(evt.DisconnectionComplete, evt.HandlerFunc(h.handleDisconnectionComplete))
-	e.HandleEvent(evt.NumberOfCompletedPkts, evt.HandlerFunc(h.handleNumberOfCompletedPkts))
-	e.HandleEvent(evt.CommandComplete, evt.HandlerFunc(c.HandleComplete))
-	e.HandleEvent(evt.CommandStatus, evt.HandlerFunc(c.HandleStatus))
+	ev.HandleEvent(event.LEMeta, event.HandlerFunc(h.handleLEMeta))
+	ev.HandleEvent(event.DisconnectionComplete, event.HandlerFunc(h.handleDisconnectionComplete))
+	ev.HandleEvent(event.NumberOfCompletedPkts, event.HandlerFunc(h.handleNumberOfCompletedPkts))
+	ev.HandleEvent(event.CommandComplete, event.HandlerFunc(cmd.HandleComplete))
+	ev.HandleEvent(event.CommandStatus, event.HandlerFunc(cmd.HandleStatus))
 
-	go h.mainLoop()
-	h.resetDevice()
+	go h.mainLoop(ctx)
+	h.resetDevice(ctx)
 	return h, nil
 }
 
 func (h *HCI) Close() error {
-	log.Printf("hci.Close()")
+	ctx := context.TODO()
+	logger.Debugf(ctx, "hci.Close()")
 	h.pool.Close()
 	<-h.loopDone
-	log.Printf("mainLoop exited")
+	logger.Debugf(ctx, "mainLoop exited")
 	for _, c := range h.conns {
-		log.Printf("closing connection %v", c)
+		logger.Debugf(ctx, "closing connection %v", c)
 		c.Close()
 	}
-	log.Printf("closing %v", h.d)
-	return h.d.Close()
+	logger.Debugf(ctx, "closing %v", h.device)
+	return h.device.Close()
 }
 
-func (h *HCI) SetAdvertiseEnable(en bool) error {
-	h.advmu.Lock()
+func (h *HCI) SetAdvertiseEnable(ctx context.Context, en bool) error {
+	h.advMutex.Lock()
 	h.adv = en
-	h.advmu.Unlock()
-	return h.setAdvertiseEnable(en)
+	h.advMutex.Unlock()
+	return h.setAdvertiseEnable(ctx, en)
 }
 
-func (h *HCI) setAdvertiseEnable(en bool) error {
-	h.advmu.Lock()
-	defer h.advmu.Unlock()
+func (h *HCI) setAdvertiseEnable(ctx context.Context, en bool) error {
+	h.advMutex.Lock()
+	defer h.advMutex.Unlock()
 	if en && h.adv && (len(h.conns) == h.maxConn) {
 		return nil
 	}
-	return h.c.SendAndCheckResp(
-		cmd.LESetAdvertiseEnable{
-			AdvertisingEnable: btoi(en),
-		}, []byte{0x00})
+	return h.cmd.SendAndCheckResp(ctx, cmd.LESetAdvertiseEnable{
+		AdvertisingEnable: btoi(en),
+	}, []byte{0x00})
 }
 
-func (h *HCI) SendCmdWithAdvOff(c cmd.CmdParam) error {
-	h.setAdvertiseEnable(false)
-	err := h.c.SendAndCheckResp(c, nil)
+func (h *HCI) SendCmdWithAdvOff(ctx context.Context, c cmd.CmdParam) error {
+	h.setAdvertiseEnable(ctx, false)
+	err := h.cmd.SendAndCheckResp(ctx, c, nil)
 	if h.adv {
-		h.setAdvertiseEnable(true)
+		h.setAdvertiseEnable(ctx, true)
 	}
 	return err
 }
 
-func (h *HCI) SetScanEnable(en bool, dup bool) error {
-	return h.c.SendAndCheckResp(
+func (h *HCI) SetScanEnable(ctx context.Context, en bool, dup bool) error {
+	return h.cmd.SendAndCheckResp(
+		ctx,
 		cmd.LESetScanEnable{
 			LEScanEnable:     btoi(en),
 			FilterDuplicates: btoi(!dup),
 		}, []byte{0x00})
 }
 
-func (h *HCI) Connect(pd *PlatData) error {
-	h.c.Send(
-		cmd.LECreateConn{
-			LEScanInterval:        0x0004,         // N x 0.625ms
-			LEScanWindow:          0x0004,         // N x 0.625ms
-			InitiatorFilterPolicy: 0x00,           // white list not used
-			PeerAddressType:       pd.AddressType, // public or random
-			PeerAddress:           pd.Address,     //
-			OwnAddressType:        0x00,           // public
-			ConnIntervalMin:       0x0006,         // N x 0.125ms
-			ConnIntervalMax:       0x0006,         // N x 0.125ms
-			ConnLatency:           0x0000,         //
-			SupervisionTimeout:    0x0048,         // N x 10ms
-			MinimumCELength:       0x0000,         // N x 0.625ms
-			MaximumCELength:       0x0000,         // N x 0.625ms
-		})
+func (h *HCI) Connect(ctx context.Context, pd *PlatData) error {
+	h.cmd.Send(ctx, cmd.LECreateConn{
+		LEScanInterval:        0x0004,         // N x 0.625ms
+		LEScanWindow:          0x0004,         // N x 0.625ms
+		InitiatorFilterPolicy: 0x00,           // allow list not used
+		PeerAddressType:       pd.AddressType, // public or random
+		PeerAddress:           pd.Address,     //
+		OwnAddressType:        0x00,           // public
+		ConnIntervalMin:       0x0006,         // N x 0.125ms
+		ConnIntervalMax:       0x0006,         // N x 0.125ms
+		ConnLatency:           0x0000,         //
+		SupervisionTimeout:    0x0048,         // N x 10ms
+		MinimumCELength:       0x0000,         // N x 0.625ms
+		MaximumCELength:       0x0000,         // N x 0.625ms
+	})
 	return nil
 }
 
@@ -167,8 +169,8 @@ func (h *HCI) CancelConnection(pd *PlatData) error {
 	return nil
 }
 
-func (h *HCI) SendRawCommand(c cmd.CmdParam) ([]byte, error) {
-	return h.c.Send(c)
+func (h *HCI) SendRawCommand(ctx context.Context, c cmd.CmdParam) ([]byte, error) {
+	return h.cmd.Send(ctx, c)
 }
 
 func btoi(b bool) uint8 {
@@ -178,73 +180,76 @@ func btoi(b bool) uint8 {
 	return 0
 }
 
-func (h *HCI) mainLoop() {
-	log.Printf("hci.mainLoop started")
+func (h *HCI) mainLoop(ctx context.Context) {
+	logger.Debugf(ctx, "hci.mainLoop started")
 	defer func() {
 		h.loopDone <- true
 	}()
 
 	for {
-		// log.Printf("hci.mainLoop pool.Get")
+		// logger.Debugf(ctx, "hci.mainLoop pool.Get")
 		b := h.pool.Get()
 		if b == nil {
-			log.Printf("got nil buffer, breaking mainLoop")
+			logger.Debugf(ctx, "got nil buffer, breaking mainLoop")
 			break
 		}
-		// log.Printf("hci.mainLoop Read(%d)", len(b))
-		n, err := h.d.Read(b)
+		// logger.Debugf(ctx, "hci.mainLoop Read(%d)", len(b))
+		n, err := h.device.Read(b)
 		if err != nil {
 			if err == unix.EINTR {
 				continue
 			}
 
-			log.Printf("mainloop err: %v", err)
+			logger.Debugf(ctx, "mainLoop err: %v", err)
 			return
 		}
 		if n > 0 {
-			// log.Printf("hci.mainLoop -> handlePacket")
-			h.handlePacket(b, n)
+			// logger.Debugf(ctx, "hci.mainLoop -> handlePacket")
+			h.handlePacket(ctx, b, n)
 		}
 	}
-	log.Printf("hci.mainLoop stopped")
+	logger.Debugf(ctx, "hci.mainLoop stopped")
 }
 
-func (h *HCI) handlePacket(buf []byte, n int) {
+var ErrVendorNotSupported = fmt.Errorf("vendor packet not supported")
+
+func (h *HCI) handlePacket(ctx context.Context, buf []byte, n int) {
 	b := buf[:n]
-	t, b := packetType(b[0]), b[1:]
+	t, b := consts.PacketType(b[0]), b[1:]
 	var err error
 	handled := true
 	switch t {
-	case typCommandPkt:
+	case consts.PacketTypeCommand:
 		op := uint16(b[0]) | uint16(b[1])<<8
-		log.Printf("unmanaged cmd: opcode (%04x) [ % X ]\n", op, b)
-	case typACLDataPkt:
-		err = h.handleL2CAP(b)
-	case typSCODataPkt:
+		logger.Debugf(ctx, "unmanaged cmd: opcode (%04x) [ % X ]\n", op, b)
+	case consts.PacketTypeACLData:
+		err = h.handleL2CAP(ctx, b)
+	case consts.PacketTypeSCOData:
 		err = fmt.Errorf("SCO packet not supported")
-	case typEventPkt:
+	case consts.PacketTypeEvent:
 		handled = false
 		go func() {
-			err := h.e.Dispatch(b)
+			err := h.event.Dispatch(ctx, b)
 			if err != nil {
-				log.Printf("hci: %s, [ % X]", err, b)
+				logger.Debugf(ctx, "hci: %s, [ % X]", err, b)
 			}
 			h.pool.Put(buf)
 		}()
-	case typVendorPkt:
-		err = fmt.Errorf("Vendor packet not supported")
+	case consts.PacketTypeVendor:
+		err = ErrVendorNotSupported
 	default:
-		log.Fatalf("Unknown event: 0x%02X [ % X ]\n", t, b)
+		logger.Fatalf(ctx, "Unknown event: 0x%02X [ % X ]\n", t, b)
 	}
 	if err != nil {
-		log.Printf("hci: %s, [ % X]", err, b)
+		logger.Debugf(ctx, "hci: %s, [ % X]", err, b)
 	}
 	if handled {
 		h.pool.Put(buf)
 	}
 }
 
-func (h *HCI) resetDevice() error {
+func (h *HCI) resetDevice(ctx context.Context) error {
+	// TODO: explain all the magic numbers below:
 	seq := []cmd.CmdParam{
 		cmd.Reset{},
 		cmd.SetEventMask{EventMask: 0x3dbff807fffbffff},
@@ -254,24 +259,25 @@ func (h *HCI) resetDevice() error {
 		cmd.WriteInquiryMode{InquiryMode: 2},
 		cmd.WritePageScanType{PageScanType: 1},
 		cmd.WriteInquiryScanType{ScanType: 1},
-		cmd.WriteClassOfDevice{ClassOfDevice: [3]byte{0x40, 0x02, 0x04}},
+		cmd.WriteDeviceClass{DeviceClass: [3]byte{0x40, 0x02, 0x04}},
 		cmd.WritePageTimeout{PageTimeout: 0x2000},
 		cmd.WriteDefaultLinkPolicy{DefaultLinkPolicySettings: 0x5},
 		cmd.HostBufferSize{
 			HostACLDataPacketLength:            0x1000,
 			HostSynchronousDataPacketLength:    0xff,
 			HostTotalNumACLDataPackets:         0x0014,
-			HostTotalNumSynchronousDataPackets: 0x000a},
+			HostTotalNumSynchronousDataPackets: 0x000a,
+		},
 		cmd.LESetScanParameters{
 			LEScanType:           0x01,   // [0x00]: passive, 0x01: active
 			LEScanInterval:       0x0010, // [0x10]: 0.625ms * 16
 			LEScanWindow:         0x0010, // [0x10]: 0.625ms * 16
 			OwnAddressType:       0x00,   // [0x00]: public, 0x01: random
-			ScanningFilterPolicy: 0x00,   // [0x00]: accept all, 0x01: ignore non-white-listed.
+			ScanningFilterPolicy: 0x00,   // [0x00]: accept all, 0x01: ignore non-allow-listed.
 		},
 	}
 	for _, s := range seq {
-		if err := h.c.SendAndCheckResp(s, []byte{0x00}); err != nil {
+		if err := h.cmd.SendAndCheckResp(ctx, s, []byte{0x00}); err != nil {
 			return err
 		}
 	}
@@ -283,20 +289,20 @@ func (h *HCI) handleAdvertisement(b []byte) {
 	if h.AdvertisementHandler == nil {
 		return
 	}
-	ep := &evt.LEAdvertisingReportEP{}
+	ep := &event.LEAdvertisingReportEP{}
 	if err := ep.Unmarshal(b); err != nil {
 		return
 	}
-	for i := 0; i < int(ep.NumReports); i++ {
+	for i := range int(ep.NumReports) {
 		addr := bdaddr(ep.Address[i])
 		et := ep.EventType[i]
-		connectable := et == advInd || et == advDirectInd
-		//scannable := et == advInd || et == advScanInd
+		connectable := et == consts.AdvInd || et == consts.AdvDirectInd
+		//scannable := et == consts.AdvInd || et == consts.AdvScanInd
 
-		if et == scanRsp {
-			h.plistmu.Lock()
+		if et == consts.ScanResp {
+			h.plistMutex.Lock()
 			pd, ok := h.plist[addr]
-			h.plistmu.Unlock()
+			h.plistMutex.Unlock()
 			if ok {
 				pd.Data = append(pd.Data, ep.Data[i]...)
 				h.AdvertisementHandler(pd)
@@ -311,44 +317,44 @@ func (h *HCI) handleAdvertisement(b []byte) {
 			Connectable: connectable,
 			RSSI:        ep.RSSI[i],
 		}
-		h.plistmu.Lock()
+		h.plistMutex.Lock()
 		h.plist[addr] = pd
-		h.plistmu.Unlock()
+		h.plistMutex.Unlock()
 
 		h.AdvertisementHandler(pd)
 	}
 }
 
-func (h *HCI) handleNumberOfCompletedPkts(b []byte) error {
-	ep := &evt.NumberOfCompletedPktsEP{}
+func (h *HCI) handleNumberOfCompletedPkts(ctx context.Context, b []byte) error {
+	ep := &event.NumberOfCompletedPktsEP{}
 	if err := ep.Unmarshal(b); err != nil {
 		return err
 	}
 	for _, r := range ep.Packets {
-		for i := 0; i < int(r.NumOfCompletedPkts); i++ {
+		for range int(r.NumOfCompletedPkts) {
 			<-h.bufCnt
 		}
 	}
 	return nil
 }
 
-func (h *HCI) handleConnection(b []byte) {
+func (h *HCI) handleConnection(ctx context.Context, b []byte) {
 	defer func() {
 		if err := recover(); err != nil {
-			log.Printf("error while handling connectiont: %v", err)
+			logger.Debugf(ctx, "error while handling connection: %v", err)
 		}
 	}()
 
-	ep := &evt.LEConnectionCompleteEP{}
+	ep := &event.LEConnectionCompleteEP{}
 	if err := ep.Unmarshal(b); err != nil {
 		return // FIXME
 	}
 	hh := ep.ConnectionHandle
-	c := newConn(h, hh)
-	h.connsmu.Lock()
+	c := newConn(ctx, h, hh)
+	h.connsMutex.Lock()
 	h.conns[hh] = c
-	h.connsmu.Unlock()
-	h.setAdvertiseEnable(true)
+	h.connsMutex.Unlock()
+	h.setAdvertiseEnable(ctx, true)
 
 	// FIXME: sloppiness. This call should be called by the package user once we
 	// flesh out the support of l2cap signaling packets (CID:0x0001,0x0005)
@@ -365,93 +371,93 @@ func (h *HCI) handleConnection(b []byte) {
 		h.AcceptMasterHandler(pd)
 		return
 	}
-	h.plistmu.Lock()
+	h.plistMutex.Lock()
 	if pd := h.plist[ep.PeerAddress]; pd != nil {
-		h.plistmu.Unlock()
+		h.plistMutex.Unlock()
 		pd.Conn = c
 		h.AcceptSlaveHandler(pd)
 	} else {
-		log.Printf("HCI: can't find data for %v", ep.PeerAddress)
+		logger.Debugf(ctx, "HCI: can't find data for %v", ep.PeerAddress)
 	}
 }
 
-func (h *HCI) handleDisconnectionComplete(b []byte) error {
-	ep := &evt.DisconnectionCompleteEP{}
+func (h *HCI) handleDisconnectionComplete(ctx context.Context, b []byte) error {
+	ep := &event.DisconnectionCompleteEP{}
 	if err := ep.Unmarshal(b); err != nil {
 		return err
 	}
 	hh := ep.ConnectionHandle
-	h.connsmu.Lock()
-	defer h.connsmu.Unlock()
+	h.connsMutex.Lock()
+	defer h.connsMutex.Unlock()
 	c, found := h.conns[hh]
 	if !found {
 		// should not happen, just be cautious for now.
-		log.Printf("l2conn: disconnecting a disconnected 0x%04X connection", hh)
+		logger.Debugf(ctx, "l2conn: disconnecting a disconnected 0x%04X connection", hh)
 		return nil
 	}
 	delete(h.conns, hh)
-	close(c.aclc)
-	h.setAdvertiseEnable(true)
+	close(c.aclCh)
+	h.setAdvertiseEnable(ctx, true)
 	return nil
 }
 
-func (h *HCI) handleLTKRequest(b []byte) {
-	ep := &evt.LELTKRequestEP{}
+func (h *HCI) handleLTKRequest(ctx context.Context, b []byte) {
+	ep := &event.LELTKRequestEP{}
 	if err := ep.Unmarshal(b); err != nil {
-		log.Printf("ltkrequest: error, parsing request")
+		logger.Debugf(ctx, "ltkrequest: error, parsing request")
 		return
 	}
 	hh := ep.ConnectionHandle
-	h.connsmu.Lock()
-	defer h.connsmu.Unlock()
+	h.connsMutex.Lock()
+	defer h.connsMutex.Unlock()
 	_, found := h.conns[hh]
 	if !found {
 		// should not happen, just be cautious for now.
-		log.Printf("ltkrequest: error, connection 0x%04X probably expired", hh)
+		logger.Debugf(ctx, "ltkrequest: error, connection 0x%04X probably expired", hh)
 		return
 	}
-	h.c.Send(cmd.LELTKNegReply{ConnectionHandle: hh})
+	h.cmd.Send(ctx, cmd.LELTKNegReply{ConnectionHandle: hh})
 	// TODO: implement proper key management
 }
 
-func (h *HCI) handleLEMeta(b []byte) error {
+func (h *HCI) handleLEMeta(ctx context.Context, b []byte) error {
 	defer func() {
 		if err := recover(); err != nil {
-			log.Printf("error while handling meta: %v", err)
+			logger.Debugf(ctx, "error while handling meta: %v", err)
 		}
 	}()
 
-	code := evt.LEEventCode(b[0])
+	code := event.LEEventCode(b[0])
 	switch code {
-	case evt.LEConnectionComplete:
-		go h.handleConnection(b)
-	case evt.LEConnectionUpdateComplete:
+	case event.LEConnectionComplete:
+		go h.handleConnection(ctx, b)
+	case event.LEConnectionUpdateComplete:
 		// anything to do here?
-	case evt.LEAdvertisingReport:
+	case event.LEAdvertisingReport:
 		go h.handleAdvertisement(b)
-	// case evt.LEReadRemoteUsedFeaturesComplete:
-	case evt.LELTKRequest:
-		go h.handleLTKRequest(b)
-	// case evt.LERemoteConnectionParameterRequest:
+	// case event.LEReadRemoteUsedFeaturesComplete:
+	case event.LELTKRequest:
+		go h.handleLTKRequest(ctx, b)
+	// case event.LERemoteConnectionParameterRequest:
 	default:
-		return fmt.Errorf("Unhandled LE event: 0x%02x, [ % X ]", code, b)
+		return fmt.Errorf("unhandled LE event: 0x%02x, [ % X ]", code, b)
 	}
 	return nil
 }
 
-func (h *HCI) handleL2CAP(b []byte) error {
+func (h *HCI) handleL2CAP(ctx context.Context, b []byte) error {
 	a := &aclData{}
 	if err := a.unmarshal(b); err != nil {
 		return err
 	}
-	h.connsmu.Lock()
-	defer h.connsmu.Unlock()
+	h.connsMutex.Lock()
+	defer h.connsMutex.Unlock()
 	c, found := h.conns[a.attr]
 	if !found {
 		// should not happen, just be cautious for now.
-		log.Printf("l2conn: got data for disconnected handle: 0x%04x", a.attr)
+		logger.Debugf(ctx, "l2conn: got data for disconnected handle: 0x%04x", a.attr)
 		return nil
 	}
-	c.aclc <- a
+	c.aclCh <- a
 	return nil
 }
