@@ -1,26 +1,27 @@
 package gatt
 
 import (
+	"context"
 	"sync"
 
-	"github.com/photostorm/gatt/xpc"
+	"github.com/xaionaro-go/gatt/xpc"
 )
 
 type central struct {
-	dev         *device
-	uuid        UUID
-	mtu         int
-	notifiers   map[uint16]*notifier
-	notifiersmu *sync.Mutex
+	dev            *device
+	uuid           UUID
+	mtu            int
+	notifiers      map[uint16]*notifier
+	notifiersMutex *sync.Mutex
 }
 
 func newCentral(d *device, u UUID) *central {
 	return &central{
-		dev:         d,
-		mtu:         23,
-		uuid:        u,
-		notifiers:   make(map[uint16]*notifier),
-		notifiersmu: &sync.Mutex{},
+		dev:            d,
+		mtu:            23,
+		uuid:           u,
+		notifiers:      make(map[uint16]*notifier),
+		notifiersMutex: &sync.Mutex{},
 	}
 }
 
@@ -28,10 +29,10 @@ func (c *central) ID() string   { return c.uuid.String() }
 func (c *central) Close() error { return nil }
 func (c *central) MTU() int     { return c.mtu }
 
-func (c *central) sendNotification(a *attr, b []byte) (int, error) {
+func (c *central) sendNotification(ctx context.Context, a *attr, b []byte) (int, error) {
 	data := make([]byte, len(b))
 	copy(data, b) // have to make a copy, why?
-	c.dev.sendCmd(15, xpc.Dict{
+	c.dev.sendCmd(ctx, 15, xpc.Dict{
 		// "kCBMsgArgUUIDs": [][]byte{reverse(c.uuid.b)}, // connection interrupted
 		// "kCBMsgArgUUIDs": [][]byte{c.uuid.b}, // connection interrupted
 		// "kCBMsgArgUUIDs": []xpc.UUID{xpc.UUID(reverse(c.uuid.b))},
@@ -48,21 +49,21 @@ func (c *central) sendNotification(a *attr, b []byte) (int, error) {
 	return len(b), nil
 }
 
-func (c *central) startNotify(a *attr, maxlen int) {
-	c.notifiersmu.Lock()
-	defer c.notifiersmu.Unlock()
+func (c *central) startNotify(ctx context.Context, a *attr, maxLen int) {
+	c.notifiersMutex.Lock()
+	defer c.notifiersMutex.Unlock()
 	if _, found := c.notifiers[a.h]; found {
 		return
 	}
-	n := newNotifier(c, a, maxlen)
+	n := newNotifier(c, a, maxLen)
 	c.notifiers[a.h] = n
 	char := a.pvt.(*Characteristic)
-	go char.nhandler.ServeNotify(Request{Central: c}, n)
+	go char.notifyHandler.ServeNotify(ctx, Request{Central: c}, n)
 }
 
 func (c *central) stopNotify(a *attr) {
-	c.notifiersmu.Lock()
-	defer c.notifiersmu.Unlock()
+	c.notifiersMutex.Lock()
+	defer c.notifiersMutex.Unlock()
 	if n, found := c.notifiers[a.h]; found {
 		n.stop()
 		delete(c.notifiers, a.h)
