@@ -1,8 +1,11 @@
 package gatt
 
-import "errors"
+import (
+	"context"
+	"errors"
+)
 
-var notImplemented = errors.New("not implemented")
+var errNotImplemented = errors.New("not implemented")
 
 type State int
 
@@ -29,64 +32,64 @@ func (s State) String() string {
 
 // Device defines the interface for a BLE device.
 // Since an interface can't define fields(properties). To implement the
-// callback support for cerntain events, deviceHandler is defined and
+// callback support for certain events, deviceHandler is defined and
 // implementation of Device on different platforms should embed it in
 // order to keep have keep compatible in API level.
 // Package users can use the Handler to set these handlers.
 type Device interface {
-	DevID() int
+	ID() int
 
-	Init(stateChanged func(Device, State)) error
-
-	// Advertise advertise AdvPacket
-	Advertise(a *AdvPacket) error
-
-	// AdvertiseNameAndServices advertises device name, and specified service UUIDs.
-	// It tres to fit the UUIDs in the advertising packet as much as possible.
-	// If name doesn't fit in the advertising packet, it will be put in scan response.
-	AdvertiseNameAndServices(name string, ss []UUID) error
-
-	AdvertiseNameAndIBeaconData(name string, b []byte) error
-
-	// AdvertiseIBeaconData advertise iBeacon with given manufacturer data.
-	AdvertiseIBeaconData(b []byte) error
-
-	// AdvertisingIbeacon advertises iBeacon with specified parameters.
-	AdvertiseIBeacon(u UUID, major, minor uint16, pwr int8) error
-
-	// StopAdvertising stops advertising.
-	StopAdvertising() error
-
-	// RemoveAllServices removes all services that are currently in the database.
-	RemoveAllServices() error
-
-	// Add Service add a service to database.
-	AddService(s *Service) error
-
-	// SetServices set the specified service to the database.
-	// It removes all currently added services, if any.
-	SetServices(ss []*Service) error
-
-	// Scan discovers surounding remote peripherals that have the Service UUID specified in ss.
-	// If ss is set to nil, all devices scanned are reported.
-	// dup specifies weather duplicated advertisement should be reported or not.
-	// When a remote peripheral is discovered, the PeripheralDiscovered Handler is called.
-	Scan(ss []UUID, dup bool)
-
-	// StopScanning stops scanning.
-	StopScanning()
+	Start(ctx context.Context, stateChanged func(context.Context, Device, State)) error
 
 	// Stop calls OS specific close calls
 	Stop() error
 
+	// Advertise advertise AdvPacket
+	Advertise(ctx context.Context, a *AdvPacket) error
+
+	// AdvertiseNameAndServices advertises device name, and specified service UUIDs.
+	// It tres to fit the UUIDs in the advertising packet as much as possible.
+	// If name doesn't fit in the advertising packet, it will be put in scan response.
+	AdvertiseNameAndServices(ctx context.Context, name string, ss []UUID) error
+
+	AdvertiseNameAndIBeaconData(ctx context.Context, name string, b []byte) error
+
+	// AdvertiseIBeaconData advertise iBeacon with given manufacturer data.
+	AdvertiseIBeaconData(ctx context.Context, b []byte) error
+
+	// AdvertisingIBeacon advertises iBeacon with specified parameters.
+	AdvertiseIBeacon(ctx context.Context, u UUID, major, minor uint16, pwr int8) error
+
+	// StopAdvertising stops advertising.
+	StopAdvertising(ctx context.Context) error
+
+	// RemoveAllServices removes all services that are currently in the database.
+	RemoveAllServices(ctx context.Context) error
+
+	// Add Service add a service to database.
+	AddService(ctx context.Context, s *Service) error
+
+	// SetServices set the specified service to the database.
+	// It removes all currently added services, if any.
+	SetServices(ctx context.Context, ss []*Service) error
+
+	// Scan discovers surrounding remote peripherals that have the Service UUID specified in ss.
+	// If ss is set to nil, all devices scanned are reported.
+	// dup specifies weather duplicated advertisement should be reported or not.
+	// When a remote peripheral is discovered, the PeripheralDiscovered Handler is called.
+	Scan(ctx context.Context, ss []UUID, dup bool) error
+
+	// StopScanning stops scanning.
+	StopScanning() error
+
 	// Connect connects to a remote peripheral.
-	Connect(p Peripheral)
+	Connect(ctx context.Context, p Peripheral)
 
 	// CancelConnection disconnects a remote peripheral.
-	CancelConnection(p Peripheral)
+	CancelConnection(ctx context.Context, p Peripheral)
 
 	// Handle registers the specified handlers.
-	Handle(h ...Handler)
+	Handle(ctx context.Context, h ...Handler)
 
 	// Option sets the options specified.
 	Option(o ...Option) error
@@ -95,22 +98,22 @@ type Device interface {
 // deviceHandler is the handlers(callbacks) of the Device.
 type deviceHandler struct {
 	// stateChanged is called when the device states changes.
-	stateChanged func(d Device, s State)
+	stateChanged func(ctx context.Context, d Device, s State)
 
 	// connect is called when a remote central device connects to the device.
-	centralConnected func(c Central)
+	centralConnected func(ctx context.Context, c Central)
 
 	// disconnect is called when a remote central device disconnects to the device.
-	centralDisconnected func(c Central)
+	centralDisconnected func(ctx context.Context, c Central)
 
 	// peripheralDiscovered is called when a remote peripheral device is found during scan procedure.
-	peripheralDiscovered func(p Peripheral, a *Advertisement, rssi int)
+	peripheralDiscovered func(ctx context.Context, p Peripheral, a *Advertisement, rssi int)
 
-	// peripheralConnected is called when a remote peripheral is conneted.
-	peripheralConnected func(p Peripheral, err error)
+	// peripheralConnected is called when a remote peripheral is connected.
+	peripheralConnected func(ctx context.Context, p Peripheral, err error)
 
-	// peripheralConnected is called when a remote peripheral is disconneted.
-	peripheralDisconnected func(p Peripheral, err error)
+	// peripheralConnected is called when a remote peripheral is disconnected.
+	peripheralDisconnected func(ctx context.Context, p Peripheral, err error)
 }
 
 func getDeviceHandler(d Device) *deviceHandler {
@@ -126,42 +129,42 @@ func getDeviceHandler(d Device) *deviceHandler {
 
 // A Handler is a self-referential function, which registers the options specified.
 // See http://commandcenter.blogspot.com.au/2014/01/self-referential-functions-and-design.html for more discussion.
-type Handler func(Device)
+type Handler func(context.Context, Device)
 
 // Handle registers the specified handlers.
-func (d *device) Handle(hh ...Handler) {
+func (d *device) Handle(ctx context.Context, hh ...Handler) {
 	for _, h := range hh {
-		h(d)
+		h(ctx, d)
 	}
 }
 
 // CentralConnected returns a Handler, which sets the specified function to be called when a device connects to the server.
-func CentralConnected(f func(Central)) Handler {
-	return func(d Device) { getDeviceHandler(d).centralConnected = f }
+func CentralConnected(f func(context.Context, Central)) Handler {
+	return func(ctx context.Context, d Device) { getDeviceHandler(d).centralConnected = f }
 }
 
 // CentralDisconnected returns a Handler, which sets the specified function to be called when a device disconnects from the server.
-func CentralDisconnected(f func(Central)) Handler {
-	return func(d Device) { getDeviceHandler(d).centralDisconnected = f }
+func CentralDisconnected(f func(context.Context, Central)) Handler {
+	return func(ctx context.Context, d Device) { getDeviceHandler(d).centralDisconnected = f }
 }
 
 // PeripheralDiscovered returns a Handler, which sets the specified function to be called when a remote peripheral device is found during scan procedure.
-func PeripheralDiscovered(f func(Peripheral, *Advertisement, int)) Handler {
-	return func(d Device) { getDeviceHandler(d).peripheralDiscovered = f }
+func PeripheralDiscovered(f func(context.Context, Peripheral, *Advertisement, int)) Handler {
+	return func(ctx context.Context, d Device) { getDeviceHandler(d).peripheralDiscovered = f }
 }
 
 // PeripheralConnected returns a Handler, which sets the specified function to be called when a remote peripheral device connects.
-func PeripheralConnected(f func(Peripheral, error)) Handler {
-	return func(d Device) { getDeviceHandler(d).peripheralConnected = f }
+func PeripheralConnected(f func(context.Context, Peripheral, error)) Handler {
+	return func(ctx context.Context, d Device) { getDeviceHandler(d).peripheralConnected = f }
 }
 
 // PeripheralDisconnected returns a Handler, which sets the specified function to be called when a remote peripheral device disconnects.
-func PeripheralDisconnected(f func(Peripheral, error)) Handler {
-	return func(d Device) { getDeviceHandler(d).peripheralDisconnected = f }
+func PeripheralDisconnected(f func(context.Context, Peripheral, error)) Handler {
+	return func(ctx context.Context, d Device) { getDeviceHandler(d).peripheralDisconnected = f }
 }
 
 // An Option is a self-referential function, which sets the option specified.
-// Most Options are platform-specific, which gives more fine-grained control over the device at a cost of losing portibility.
+// Most Options are platform-specific, which gives more fine-grained control over the device at a cost of losing portability.
 // See http://commandcenter.blogspot.com.au/2014/01/self-referential-functions-and-design.html for more discussion.
 type Option func(Device) error
 
