@@ -1,22 +1,27 @@
 package gatt
 
 import (
+	"context"
 	"encoding/binary"
 	"net"
 
-	"github.com/photostorm/gatt/linux"
-	"github.com/photostorm/gatt/linux/cmd"
+	"github.com/xaionaro-go/ctxflow"
+	"github.com/xaionaro-go/gatt/linux"
+	"github.com/xaionaro-go/gatt/linux/cmd"
 )
 
 type device struct {
 	deviceHandler
 
+	mainLoop ctxflow.StartStopper[ctxflow.StartStopperBackendFuncs]
+	scanner  ctxflow.StartStopper[ctxflow.StartStopperBackendFuncs]
+
 	hci   *linux.HCI
 	state State
 
 	// All the following fields are only used peripheralManager (server) implementation.
-	svcs  []*Service
-	attrs *attrRange
+	services []*Service
+	attrs    *attrRange
 
 	devID   int
 	chkLE   bool
@@ -28,7 +33,7 @@ type device struct {
 	scanParam *cmd.LESetScanParameters
 }
 
-func NewDevice(opts ...Option) (Device, error) {
+func NewDevice(ctx context.Context, opts ...Option) (Device, error) {
 	d := &device{
 		maxConn: 1,    // Support 1 connection at a time.
 		devID:   -1,   // Find an available HCI device.
@@ -49,12 +54,25 @@ func NewDevice(opts ...Option) (Device, error) {
 			LEScanInterval:       0x0010, // [0x10]: 0.625ms * 16
 			LEScanWindow:         0x0010, // [0x10]: 0.625ms * 16
 			OwnAddressType:       0x00,   // [0x00]: public, 0x01: random
-			ScanningFilterPolicy: 0x00,   // [0x00]: accept all, 0x01: ignore non-white-listed.
+			ScanningFilterPolicy: 0x00,   // [0x00]: accept all, 0x01: ignore non-allow-listed.
+		},
+	}
+
+	d.mainLoop = ctxflow.StartStopper[ctxflow.StartStopperBackendFuncs]{
+		StartStopper: ctxflow.StartStopperBackendFuncs{
+			StartFunc: d.doStartMainLoop,
+			StopFunc:  d.doStopMainLoop,
+		},
+	}
+	d.scanner = ctxflow.StartStopper[ctxflow.StartStopperBackendFuncs]{
+		StartStopper: ctxflow.StartStopperBackendFuncs{
+			StartFunc: d.doStartScanning,
+			StopFunc:  d.doStopScanning,
 		},
 	}
 
 	d.Option(opts...)
-	h, err := linux.NewHCI(d.devID, d.chkLE, d.maxConn)
+	h, err := linux.NewHCI(ctx, d.devID, d.chkLE, d.maxConn)
 	if err != nil {
 		return nil, err
 	}
@@ -63,94 +81,104 @@ func NewDevice(opts ...Option) (Device, error) {
 	return d, nil
 }
 
-func (d *device) Init(f func(Device, State)) error {
+func (d *device) Start(ctx context.Context, f func(context.Context, Device, State)) error {
+	return d.mainLoop.Start(ctx, f)
+}
+
+func (d *device) Stop() error {
+	return d.mainLoop.Stop()
+}
+
+func (d *device) doStartMainLoop(ctx context.Context, args ...any) error {
+	f := args[0].(func(context.Context, Device, State))
+
 	d.hci.AcceptMasterHandler = func(pd *linux.PlatData) {
-		a := pd.Address
-		c := newCentral(d.attrs, net.HardwareAddr([]byte{a[5], a[4], a[3], a[2], a[1], a[0]}), pd.Conn)
+		addr := pd.Address
+		central := newCentral(d.attrs, net.HardwareAddr([]byte{addr[5], addr[4], addr[3], addr[2], addr[1], addr[0]}), pd.Conn)
 		if d.centralConnected != nil {
-			d.centralConnected(c)
+			d.centralConnected(ctx, central)
 		}
-		c.loop()
+		central.loop(ctx)
 		if d.centralDisconnected != nil {
-			d.centralDisconnected(c)
+			d.centralDisconnected(ctx, central)
 		}
 	}
 	d.hci.AcceptSlaveHandler = func(pd *linux.PlatData) {
-		p := &peripheral{
-			d:     d,
-			pd:    pd,
-			l2c:   pd.Conn,
-			reqc:  make(chan message),
-			quitc: make(chan struct{}),
-			sub:   newSubscriber(),
+		periph := &peripheral{
+			d:      d,
+			pd:     pd,
+			l2c:    pd.Conn,
+			reqCh:  make(chan message),
+			quitCh: make(chan struct{}),
+			sub:    newSubscriber(),
 		}
 		if d.peripheralConnected != nil {
-			go d.peripheralConnected(p, nil)
+			go d.peripheralConnected(ctx, periph, nil)
 		}
-		p.loop()
+		periph.loop(ctx)
 		if d.peripheralDisconnected != nil {
-			d.peripheralDisconnected(p, nil)
+			d.peripheralDisconnected(ctx, periph, nil)
 		}
 	}
 	d.hci.AdvertisementHandler = func(pd *linux.PlatData) {
-		a := &Advertisement{}
-		a.unmarshall(pd.Data)
-		a.Connectable = pd.Connectable
-		p := &peripheral{pd: pd, d: d}
+		adv := &Advertisement{}
+		adv.unmarshall(pd.Data)
+		adv.Connectable = pd.Connectable
+		periph := &peripheral{pd: pd, d: d}
 		if d.peripheralDiscovered != nil {
-			pd.Name = a.LocalName
-			d.peripheralDiscovered(p, a, int(pd.RSSI))
+			pd.Name = adv.LocalName
+			d.peripheralDiscovered(ctx, periph, adv, int(pd.RSSI))
 		}
 	}
 	d.state = StatePoweredOn
 	d.stateChanged = f
-	go d.stateChanged(d, d.state)
+	go d.stateChanged(ctx, d, d.state)
 	return nil
 }
 
-func (d *device) DevID() int {
-	return d.devID
-}
-
-func (d *device) Stop() error {
+func (d *device) doStopMainLoop(ctx context.Context) error {
 	d.state = StatePoweredOff
-	defer d.stateChanged(d, d.state)
+	defer d.stateChanged(ctx, d, d.state)
 	return d.hci.Close()
 }
 
-func (d *device) AddService(s *Service) error {
-	d.svcs = append(d.svcs, s)
-	d.attrs = generateAttributes(d.svcs, uint16(1)) // ble attrs start at 1
+func (d *device) ID() int {
+	return d.devID
+}
+
+func (d *device) AddService(ctx context.Context, s *Service) error {
+	d.services = append(d.services, s)
+	d.attrs = generateAttributes(ctx, d.services, uint16(1)) // ble attrs start at 1
 	return nil
 }
 
-func (d *device) RemoveAllServices() error {
-	d.svcs = nil
+func (d *device) RemoveAllServices(ctx context.Context) error {
+	d.services = nil
 	d.attrs = nil
 	return nil
 }
 
-func (d *device) SetServices(s []*Service) error {
-	d.RemoveAllServices()
-	d.svcs = append(d.svcs, s...)
-	d.attrs = generateAttributes(d.svcs, uint16(1)) // ble attrs start at 1
+func (d *device) SetServices(ctx context.Context, s []*Service) error {
+	d.RemoveAllServices(ctx)
+	d.services = append(d.services, s...)
+	d.attrs = generateAttributes(ctx, d.services, uint16(1)) // ble attrs start at 1
 	return nil
 }
 
-func (d *device) Advertise(a *AdvPacket) error {
+func (d *device) Advertise(ctx context.Context, a *AdvPacket) error {
 	d.advData = &cmd.LESetAdvertisingData{
 		AdvertisingDataLength: uint8(a.Len()),
 		AdvertisingData:       a.Bytes(),
 	}
 
-	if err := d.update(); err != nil {
+	if err := d.update(ctx); err != nil {
 		return err
 	}
 
-	return d.hci.SetAdvertiseEnable(true)
+	return d.hci.SetAdvertiseEnable(ctx, true)
 }
 
-func (d *device) AdvertiseNameAndServices(name string, uu []UUID) error {
+func (d *device) AdvertiseNameAndServices(ctx context.Context, name string, uu []UUID) error {
 	a := &AdvPacket{}
 	a.AppendFlags(flagGeneralDiscoverable | flagLEOnly)
 	a.AppendUUIDFit(uu)
@@ -167,10 +195,10 @@ func (d *device) AdvertiseNameAndServices(name string, uu []UUID) error {
 		}
 	}
 
-	return d.Advertise(a)
+	return d.Advertise(ctx, a)
 }
 
-func (d *device) AdvertiseNameAndIBeaconData(name string, b []byte) error {
+func (d *device) AdvertiseNameAndIBeaconData(ctx context.Context, name string, b []byte) error {
 	a := &AdvPacket{}
 	a.AppendFlags(flagGeneralDiscoverable | flagLEOnly)
 	a.AppendManufacturerData(0x004C, b)
@@ -187,18 +215,18 @@ func (d *device) AdvertiseNameAndIBeaconData(name string, b []byte) error {
 		}
 	}
 
-	return d.Advertise(a)
+	return d.Advertise(ctx, a)
 }
 
-func (d *device) AdvertiseIBeaconData(b []byte) error {
+func (d *device) AdvertiseIBeaconData(ctx context.Context, b []byte) error {
 	a := &AdvPacket{}
 	a.AppendFlags(flagGeneralDiscoverable | flagLEOnly)
 	a.AppendManufacturerData(0x004C, b)
 
-	return d.Advertise(a)
+	return d.Advertise(ctx, a)
 }
 
-func (d *device) AdvertiseIBeacon(u UUID, major, minor uint16, pwr int8) error {
+func (d *device) AdvertiseIBeacon(ctx context.Context, u UUID, major, minor uint16, pwr int8) error {
 	b := make([]byte, 23)
 	b[0] = 0x02                               // Data type: iBeacon
 	b[1] = 0x15                               // Data length: 21 bytes
@@ -206,50 +234,58 @@ func (d *device) AdvertiseIBeacon(u UUID, major, minor uint16, pwr int8) error {
 	binary.BigEndian.PutUint16(b[18:], major) // Big endian
 	binary.BigEndian.PutUint16(b[20:], minor) // Big endian
 	b[22] = uint8(pwr)                        // Measured Tx Power
-	return d.AdvertiseIBeaconData(b)
+	return d.AdvertiseIBeaconData(ctx, b)
 }
 
-func (d *device) StopAdvertising() error {
-	return d.hci.SetAdvertiseEnable(false)
+func (d *device) StopAdvertising(ctx context.Context) error {
+	return d.hci.SetAdvertiseEnable(ctx, false)
 }
 
-func (d *device) Scan(ss []UUID, dup bool) {
-	// TODO: filter
-	d.hci.SetScanEnable(true, dup)
+func (d *device) Scan(ctx context.Context, services []UUID, dup bool) error {
+	return d.scanner.Start(ctx, services, dup)
 }
 
-func (d *device) StopScanning() {
-	d.hci.SetScanEnable(false, true)
+func (d *device) StopScanning() error {
+	return d.scanner.Stop()
 }
 
-func (d *device) Connect(p Peripheral) {
-	d.hci.Connect(p.(*peripheral).pd)
+func (d *device) doStartScanning(ctx context.Context, args ...any) error {
+	_, dup := args[0].([]UUID), args[1].(bool)
+	return d.hci.SetScanEnable(ctx, true, dup)
 }
 
-func (d *device) CancelConnection(p Peripheral) {
+func (d *device) doStopScanning(ctx context.Context) error {
+	return d.hci.SetScanEnable(ctx, false, true)
+}
+
+func (d *device) Connect(ctx context.Context, p Peripheral) {
+	d.hci.Connect(ctx, p.(*peripheral).pd)
+}
+
+func (d *device) CancelConnection(ctx context.Context, p Peripheral) {
 	d.hci.CancelConnection(p.(*peripheral).pd)
 }
 
-func (d *device) SendHCIRawCommand(c cmd.CmdParam) ([]byte, error) {
-	return d.hci.SendRawCommand(c)
+func (d *device) SendHCIRawCommand(ctx context.Context, c cmd.CmdParam) ([]byte, error) {
+	return d.hci.SendRawCommand(ctx, c)
 }
 
 // Flush pending advertising settings to the device.
-func (d *device) update() error {
+func (d *device) update(ctx context.Context) error {
 	if d.advParam != nil {
-		if err := d.hci.SendCmdWithAdvOff(d.advParam); err != nil {
+		if err := d.hci.SendCmdWithAdvOff(ctx, d.advParam); err != nil {
 			return err
 		}
 		d.advParam = nil
 	}
 	if d.scanResp != nil {
-		if err := d.hci.SendCmdWithAdvOff(d.scanResp); err != nil {
+		if err := d.hci.SendCmdWithAdvOff(ctx, d.scanResp); err != nil {
 			return err
 		}
 		d.scanResp = nil
 	}
 	if d.advData != nil {
-		if err := d.hci.SendCmdWithAdvOff(d.advData); err != nil {
+		if err := d.hci.SendCmdWithAdvOff(ctx, d.advData); err != nil {
 			return err
 		}
 		d.advData = nil
