@@ -1,30 +1,31 @@
 package service
 
 import (
+	"context"
 	"fmt"
-	"log"
+
 	"time"
 
-	"github.com/photostorm/gatt"
+	"github.com/xaionaro-go/gatt"
 )
 
 func NewCountService() *gatt.Service {
 	n := 0
 	s := gatt.NewService(gatt.MustParseUUID("09fc95c0-c111-11e3-9904-0002a5d5c51b"))
 	s.AddCharacteristic(gatt.MustParseUUID("11fac9e0-c111-11e3-9246-0002a5d5c51b")).HandleReadFunc(
-		func(rsp gatt.ResponseWriter, req *gatt.ReadRequest) {
-			fmt.Fprintf(rsp, "count: %d", n)
+		func(ctx context.Context, resp gatt.ResponseWriter, req *gatt.ReadRequest) {
+			fmt.Fprintf(resp, "count: %d", n)
 			n++
 		})
 
 	s.AddCharacteristic(gatt.MustParseUUID("16fe0d80-c111-11e3-b8c8-0002a5d5c51b")).HandleWriteFunc(
-		func(r gatt.Request, data []byte) (status byte) {
-			log.Println("Wrote:", string(data))
+		func(ctx context.Context, r gatt.Request, data []byte) (status byte) {
+			fmt.Println("Wrote:", string(data))
 			return gatt.StatusSuccess
 		})
 
 	s.AddCharacteristic(gatt.MustParseUUID("1c927b50-c116-11e3-8a33-0800200c9a66")).HandleNotifyFunc(
-		func(r gatt.Request, n gatt.Notifier) {
+		func(ctx context.Context, r gatt.Request, n gatt.Notifier) {
 			cnt := 0
 			for !n.Done() {
 				fmt.Fprintf(n, "Count: %d", cnt)
