@@ -1,2 +1,2 @@
-// Package service provides a collection of sample services for demostrating purpose.
+// Package service provides a collection of sample services for demonstrating purpose.
 package service
