@@ -1,6 +1,6 @@
 package service
 
-import "github.com/photostorm/gatt"
+import "github.com/xaionaro-go/gatt"
 
 var (
 	attrGAPUUID = gatt.UUID16(0x1800)
@@ -9,7 +9,7 @@ var (
 	attrAppearanceUUID        = gatt.UUID16(0x2A01)
 	attrPeripheralPrivacyUUID = gatt.UUID16(0x2A02)
 	attrReconnectionAddrUUID  = gatt.UUID16(0x2A03)
-	attrPeferredParamsUUID    = gatt.UUID16(0x2A04)
+	attrPreferredParamsUUID   = gatt.UUID16(0x2A04)
 )
 
 // https://developer.bluetooth.org/gatt/characteristics/Pages/CharacteristicViewer.aspx?u=org.bluetooth.characteristic.gap.appearance.xml
@@ -23,6 +23,6 @@ func NewGapService(name string) *gatt.Service {
 	s.AddCharacteristic(attrAppearanceUUID).SetValue(gapCharAppearanceGenericComputer)
 	s.AddCharacteristic(attrPeripheralPrivacyUUID).SetValue([]byte{0x00})
 	s.AddCharacteristic(attrReconnectionAddrUUID).SetValue([]byte{0x00, 0x00, 0x00, 0x00, 0x00, 0x00})
-	s.AddCharacteristic(attrPeferredParamsUUID).SetValue([]byte{0x06, 0x00, 0x06, 0x00, 0x00, 0x00, 0xd0, 0x07})
+	s.AddCharacteristic(attrPreferredParamsUUID).SetValue([]byte{0x06, 0x00, 0x06, 0x00, 0x00, 0x00, 0xd0, 0x07})
 	return s
 }
