@@ -1,3 +1,4 @@
+//go:build !386
 // +build !386
 
 package socket
@@ -7,16 +8,16 @@ import (
 	"unsafe"
 )
 
-func bind(s int, addr unsafe.Pointer, addrlen _Socklen) (err error) {
-	_, _, e1 := syscall.Syscall(syscall.SYS_BIND, uintptr(s), uintptr(addr), uintptr(addrlen))
+func bind(s int, addr unsafe.Pointer, addrLen SockLen) (err error) {
+	_, _, e1 := syscall.Syscall(syscall.SYS_BIND, uintptr(s), uintptr(addr), uintptr(addrLen))
 	if e1 != 0 {
 		err = e1
 	}
 	return
 }
 
-func setsockopt(s int, level int, name int, val unsafe.Pointer, vallen uintptr) (err error) {
-	_, _, e1 := syscall.Syscall6(syscall.SYS_SETSOCKOPT, uintptr(s), uintptr(level), uintptr(name), uintptr(val), uintptr(vallen), 0)
+func setsockopt(s int, level int, name int, val unsafe.Pointer, valueLen uintptr) (err error) {
+	_, _, e1 := syscall.Syscall6(syscall.SYS_SETSOCKOPT, uintptr(s), uintptr(level), uintptr(name), uintptr(val), uintptr(valueLen), 0)
 	if e1 != 0 {
 		err = e1
 	}
